refactor: refactor Product page
