+++ src/pages/Product/ProductDescription/index.ts
@@ -1,0 +1,2 @@
+import ProductDescription from './ProductDescription';
+export { ProductDescription };
