+++ src/pages/Product/ProductDescription/productDescription.module.scss
@@ -1,0 +1,33 @@
+@import '../../../styles/common';
+
+.info {
+  width: 100%;
+
+  &__title {
+    font-size: 2rem;
+    margin-bottom: 2rem;
+  }
+
+  &__description {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    margin-bottom: 4rem;
+
+    .text {
+      &:first-child {
+        display: flex;
+        flex-direction: column;
+        gap: 2rem;
+        margin-bottom: 4rem;
+      }
+
+      &:not(:first-child) {
+        span {
+          text-decoration: underline;
+          padding-right: 1rem;
+        }
+      }
+    }
+  }
+}
