--- src/pages/Product/ProductDetails/index.ts
@@ -1,2 +1,0 @@
-import ProductDetails from './ProductDetails';
-export { ProductDetails };
