--- src/pages/Product/ProductDetails/productDetails.module.scss
@@ -1,122 +1,0 @@
-@import '../../../styles/common';
-
-.product__details {
-  max-width: 50vw;
-  width: 100%;
-  padding: 6rem 2rem 10rem 0;
-
-  transition: all 0.1s;
-
-  @media all and (max-width: 850px) {
-    max-width: 100vw;
-    padding: 2rem 2rem 12rem;
-  }
-
-  @media all and (max-width: 480px) {
-    padding: 2rem 1rem 12rem;
-  }
-}
-
-// .artist {
-//   font-size: 4rem;
-// }
-
-// .title {
-//   font-size: 4rem;
-//   margin-bottom: 6rem;
-// }
-
-// .price {
-//   display: flex;
-//   gap: 2rem;
-//   font-size: 3rem;
-//   font-weight: 600;
-//   margin-bottom: 2rem;
-
-//   @media all and (max-width: 480px) {
-//     font-size: 2.9rem;
-//   }
-
-//   &.price__double {
-//     width: 100%;
-
-//     .price__full {
-//       color: $text-grey;
-//       display: inline;
-//       position: relative;
-
-//       &::after {
-//         content: '';
-//         position: absolute;
-//         left: -3%;
-//         top: calc(50% - 2px);
-//         width: 106%;
-//         height: 4px;
-//         background-color: $primary;
-//         z-index: 10;
-//       }
-//     }
-//   }
-// }
-
-.button {
-  width: 100%;
-  margin-bottom: 9rem;
-
-  &__remove {
-    color: $dark;
-    font-weight: 600;
-    background-color: $btn-grey;
-    box-shadow: none;
-
-    &:hover {
-      color: $dark;
-      background-color: $btn-dark-grey;
-    }
-  }
-
-  @media all and (max-width: 480px) {
-    font-size: 2.9rem;
-    margin-bottom: 7rem;
-
-    &:hover {
-      font-size: 2.88rem;
-    }
-  }
-}
-
-// .info {
-//   width: 100%;
-
-//   &__title {
-//     font-size: 2rem;
-//     margin-bottom: 2rem;
-//   }
-
-//   &__description {
-//     display: flex;
-//     flex-direction: column;
-//     gap: 0.5rem;
-//     margin-bottom: 4rem;
-
-//     .text {
-//       &:first-child {
-//         display: flex;
-//         flex-direction: column;
-//         gap: 2rem;
-//         margin-bottom: 4rem;
-//       }
-
-//       &:not(:first-child) {
-//         span {
-//           text-decoration: underline;
-//           padding-right: 1rem;
-//         }
-//       }
-//     }
-//   }
-// }
-
-.error__message {
-  @include bottom-page-message($light, $error-color);
-}
